blocked, wait for transpose is done.
