+++ csrc/scheduler/registry.cpp
@@ -572,99 +572,6 @@ class TransposeScheduler : public SchedulerEntry {
   }
 };
 
-class PointWiseScheduler : public SchedulerEntry {
- public:
-  explicit PointWiseScheduler(
-      Fusion* fusion,
-      SchedulerRuntimeInfo& runtime_info,
-      HeuristicSummary* data_cache = nullptr)
-      : SchedulerEntry(ScheduleHeuristic::PointWise) {
-    computeHeuristics(fusion, runtime_info, data_cache);
-  }
-
-  static bool canScheduleCompileTime(Fusion* fusion) {
-    //   Currently using the same path as the scheduler
-    // to eliminate mismatch between canSchedule and
-    // schedule pointwise.
-    if (!hasReferenceTensorView(fusion)) {
-      scheduler_debug_utils::canScheduleRejectReason(
-          ScheduleHeuristic::PointWise, "cannot find reference tensor");
-      return false;
-    }
-
-    // Check that inputs of all select/gather-like ops are fusion inputs
-    if (registry_utils::rejectScheduleForMemoryPromotion(
-            fusion, ScheduleHeuristic::PointWise)) {
-      return false;
-    }
-
-    // Fusions handled by pointwise scheduler cannot have MmaOp.
-    if (!ir_utils::getMmaOps(fusion).empty()) {
-      scheduler_debug_utils::canScheduleRejectReason(
-          ScheduleHeuristic::PointWise, "no support for mma ops.");
-      return false;
-    }
-
-    if (!ir_utils::getViewOps(fusion).empty()) {
-      ComputeAtMap ca_map(fusion);
-      if (registry_utils::requiresForwardViewReplay(fusion, ca_map)) {
-        scheduler_debug_utils::canScheduleRejectReason(
-            ScheduleHeuristic::PointWise,
-            "Fusion requires view being reversible.");
-        return false;
-      }
-    }
-
-    auto reduction_ops = ir_utils::getReductionOps(fusion);
-
-    if (!reduction_ops.empty()) {
-      scheduler_debug_utils::canScheduleRejectReason(
-          ScheduleHeuristic::PointWise, "no support for reduction ops");
-      return false;
-    }
-
-    if (registry_utils::hasNonUniqueBcast(fusion)) {
-      scheduler_debug_utils::canScheduleRejectReason(
-          ScheduleHeuristic::PointWise,
-          "Broadcasting dimension might be broadcasting to multiple sizes.");
-      return false;
-    }
-
-    return true;
-  }
-
-  static bool canScheduleRunTime(
-      Fusion* fusion,
-      SchedulerRuntimeInfo& runtime_info,
-      HeuristicSummary* data_cache = nullptr) {
-    auto can_schedule_transpose_entry =
-        HeuristicSummaryEntry<HeuristicCompileTime::CanScheduleTranspose>(
-            data_cache, [fusion]() {
-              return std::make_unique<bool>(
-                  TransposeScheduler::canScheduleCompileTime(fusion));
-            });
-    if (can_schedule_transpose_entry.get()) {
-      auto reason =
-          getTransposeRuntimeRejectReason(fusion, data_cache, runtime_info);
-      return !reason.empty();
-    }
-
-    return true;
-  }
-
-  void schedule(Fusion* fusion) override {
-    FUSER_PERF_SCOPE("Schedule PointWise Fusion");
-    schedulePointwise(fusion, pointwiseParams());
-  }
-
-  void computeHeuristics(
-      Fusion* fusion,
-      SchedulerRuntimeInfo& runtime_info,
-      HeuristicSummary* data_cache = nullptr) {
-    params_ = getPointwiseHeuristics(fusion, runtime_info, data_cache);
-    NVF_ERROR(params_ != nullptr);
-  }
-};
 
 class MatmulScheduler : public SchedulerEntry {
  public:
